feat: add dto associations

+++ src/infrastructure/dto/coordinate.rs
@@ -1,7 +1,16 @@
-#[derive(Queryable)]
+use crate::infrastructure::schema::coordinates;
+use crate::infrastructure::dto::route::RouteDto;
+use bigdecimal::BigDecimal;
+
+
+/// 座標のdto構造体
+#[derive(Identifiable, Queryable, Insertable, Associations, Debug)]
+#[table_name = "coordinates"]
+#[primary_key(route_id, index)]
+#[belongs_to(RouteDto, foreign_key = "route_id")]
 pub struct CoordinateDto {
     pub route_id: String,
-    pub index: i32,
-    pub latitude: f64,
-    pub longitude: f64,
+    pub index: u32,
+    pub latitude: BigDecimal,
+    pub longitude: BigDecimal,
 }
